use contextmanager to do latex parse patch
